requirements.txt
numpy>1.26,<2.0
openpyxl>=3.0.10
pandas>=2.1.4
python-dateutil>=2.8.2
xlsxwriter>=3.2.0
psycopg2-binary
boto3
awslambdaric
requests
logging
